[fix] WritingPopUp  레이아웃 분리 및 Writing 폴더 재분류
